+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Features:
 * Allow `download <name>`, shortcut for `download <name>:./<name>` (downloads to current directory, keep name)
 * Allow `download --all`
 * Add `--input` and `--output` to showfiles
+* Implement `vagrant suspend` command
 
 1.0.3 (2015-12-02)
 ------------------